+++ tests/autologging/test_autologging_safety_unit.py
@@ -1,9 +1,7 @@
 import abc
-import asyncio
 import copy
 import inspect
 from collections import namedtuple
-from concurrent.futures import ThreadPoolExecutor
 from contextlib import nullcontext as does_not_raise
 from unittest import mock
 
@@ -28,10 +26,10 @@ from mlflow.utils.autologging_utils.safety import (
     _AutologgingSessionManager,
     _validate_args,
     _validate_autologging_run,
-    update_wrapper_extended,
 )
 from mlflow.utils.mlflow_tags import MLFLOW_AUTOLOGGING
 
+from tests.autologging.async_helper import asyncify, run_sync_or_sync
 from tests.autologging.fixtures import (
     patch_destination,  # noqa: F401
     test_mode_off,
@@ -167,51 +165,13 @@ def test_is_testing_respects_environment_variable(monkeypatch):
     assert is_testing()
 
 
-def _test_async(is_async):
-    """
-    Decorator that converts a function to an async function if `is_async` is True. This is useful
-    for testing purposes, where we want to test both synchronous and asynchronous code paths.
-    """
-
-    def decorator(fn):
-        if is_async:
-
-            async def async_fn(*args, **kwargs):
-                if args and inspect.iscoroutinefunction(args[0]):
-                    original = args[0]
-
-                    def wrapped_original(*og_args, **og_kwargs):
-                        # Run the original async function in a separate thread. This is a workaround
-                        # for the fact that we cannot use asyncio.run here because an event loop is
-                        # already running in the main thread.
-                        with ThreadPoolExecutor() as executor:
-                            future = executor.submit(asyncio.run, original(*og_args, **og_kwargs))
-                            return future.result()
-
-                    args = (update_wrapper_extended(wrapped_original, original), *args[1:])
-                return fn(*args, **kwargs)
-
-            return update_wrapper_extended(async_fn, fn)
-        else:
-            return fn
-
-    return decorator
-
-
-def _run_sync_or_sync(fn, *args, **kwargs):
-    if inspect.iscoroutinefunction(fn):
-        return asyncio.run(fn(*args, **kwargs))
-    else:
-        return fn(*args, **kwargs)
-
-
 def test_safe_patch_forwards_expected_arguments_to_function_based_patch_implementation(
     patch_destination, test_autologging_integration
 ):
     foo_val = None
     bar_val = None
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, foo, bar=10):
         nonlocal foo_val
         nonlocal bar_val
@@ -219,7 +179,7 @@ def test_safe_patch_forwards_expected_arguments_to_function_based_patch_implemen
         bar_val = bar
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
-    _run_sync_or_sync(patch_destination.fn, 7, bar=11)
+    run_sync_or_sync(patch_destination.fn, 7, bar=11)
     assert foo_val == 7
     assert bar_val == 11
 
@@ -227,7 +187,7 @@ def test_safe_patch_forwards_expected_arguments_to_function_based_patch_implemen
 def test_safe_patch_provides_expected_original_function(
     test_autologging_integration, patch_destination
 ):
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def original_fn(foo, bar=10):
         return {
             "foo": foo,
@@ -236,13 +196,13 @@ def test_safe_patch_provides_expected_original_function(
 
     patch_destination.fn = original_fn
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, foo, bar):
         return original(foo + 1, bar + 2)
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
 
-    assert _run_sync_or_sync(patch_destination.fn, 1, 2) == {"foo": 2, "bar": 4}
+    assert run_sync_or_sync(patch_destination.fn, 1, 2) == {"foo": 2, "bar": 4}
 
 
 def test_safe_patch_propagates_exceptions_raised_from_original_function(
@@ -250,7 +210,7 @@ def test_safe_patch_propagates_exceptions_raised_from_original_function(
 ):
     exc_to_throw = Exception("Bad original function")
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def original(*args, **kwargs):
         raise exc_to_throw
 
@@ -258,7 +218,7 @@ def test_safe_patch_propagates_exceptions_raised_from_original_function(
 
     patch_impl_called = False
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal patch_impl_called
         patch_impl_called = True
@@ -267,7 +227,7 @@ def test_safe_patch_propagates_exceptions_raised_from_original_function(
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
 
     with pytest.raises(Exception, match=str(exc_to_throw)) as exc:
-        _run_sync_or_sync(patch_destination.fn)
+        run_sync_or_sync(patch_destination.fn)
 
     assert exc.value == exc_to_throw
     assert patch_impl_called
@@ -278,15 +238,15 @@ def test_safe_patch_logs_exceptions_raised_outside_of_original_function_as_warni
 ):
     exc_to_throw = Exception("Bad patch implementation")
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         raise exc_to_throw
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
     with mock.patch("mlflow.utils.autologging_utils._logger.warning") as logger_mock:
-        assert _run_sync_or_sync(patch_destination.fn) == PATCH_DESTINATION_FN_DEFAULT_RESULT
+        assert run_sync_or_sync(patch_destination.fn) == PATCH_DESTINATION_FN_DEFAULT_RESULT
         assert logger_mock.call_count == 1
-        expected_warning = "Encountered unexpected error during {} autologging {}".format(
+        expected_warning = "Encountered unexpected error during {} autologging: {}".format(
             test_autologging_integration, exc_to_throw
         )
         assert logger_mock.call_args[0][0] == expected_warning
@@ -298,13 +258,13 @@ def test_safe_patch_propagates_exceptions_raised_outside_of_original_function_in
 ):
     exc_to_throw = Exception("Bad patch implementation")
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         raise exc_to_throw
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
     with pytest.raises(Exception, match=str(exc_to_throw)) as exc:
-        _run_sync_or_sync(patch_destination.fn)
+        run_sync_or_sync(patch_destination.fn)
 
     assert exc.value == exc_to_throw
 
@@ -314,14 +274,14 @@ def test_safe_patch_calls_original_function_when_patch_preamble_throws(
 ):
     patch_impl_called = False
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal patch_impl_called
         patch_impl_called = True
         raise Exception("Bad patch preamble")
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
-    assert _run_sync_or_sync(patch_destination.fn) == PATCH_DESTINATION_FN_DEFAULT_RESULT
+    assert run_sync_or_sync(patch_destination.fn) == PATCH_DESTINATION_FN_DEFAULT_RESULT
     assert patch_destination.fn_call_count == 1
     assert patch_impl_called
 
@@ -331,7 +291,7 @@ def test_safe_patch_returns_original_result_without_second_call_when_patch_posta
 ):
     patch_impl_called = False
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal patch_impl_called
         patch_impl_called = True
@@ -339,7 +299,7 @@ def test_safe_patch_returns_original_result_without_second_call_when_patch_posta
         raise Exception("Bad patch postamble")
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
-    assert _run_sync_or_sync(patch_destination.fn) == PATCH_DESTINATION_FN_DEFAULT_RESULT
+    assert run_sync_or_sync(patch_destination.fn) == PATCH_DESTINATION_FN_DEFAULT_RESULT
     assert patch_destination.fn_call_count == 1
     assert patch_impl_called
 
@@ -349,7 +309,7 @@ def test_safe_patch_respects_disable_flag(patch_destination):
 
     @autologging_integration("test_respects_disable")
     def autolog(disable=False, silent=False):
-        @_test_async(patch_destination.is_async)
+        @asyncify(patch_destination.is_async)
         def patch_impl(original, *args, **kwargs):
             nonlocal patch_impl_call_count
             patch_impl_call_count += 1
@@ -358,11 +318,11 @@ def test_safe_patch_respects_disable_flag(patch_destination):
         safe_patch("test_respects_disable", patch_destination, "fn", patch_impl)
 
     autolog(disable=False)
-    _run_sync_or_sync(patch_destination.fn)
+    run_sync_or_sync(patch_destination.fn)
     assert patch_impl_call_count == 1
 
     autolog(disable=True)
-    _run_sync_or_sync(patch_destination.fn)
+    run_sync_or_sync(patch_destination.fn)
     assert patch_impl_call_count == 1
 
 
@@ -371,14 +331,14 @@ def test_safe_patch_returns_original_result_and_ignores_patch_return_value(
 ):
     patch_impl_called = False
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal patch_impl_called
         patch_impl_called = True
         return 10
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
-    assert _run_sync_or_sync(patch_destination.fn) == PATCH_DESTINATION_FN_DEFAULT_RESULT
+    assert run_sync_or_sync(patch_destination.fn) == PATCH_DESTINATION_FN_DEFAULT_RESULT
     assert patch_destination.fn_call_count == 1
     assert patch_impl_called
 
@@ -387,7 +347,7 @@ def test_safe_patch_returns_original_result_and_ignores_patch_return_value(
 def test_safe_patch_validates_arguments_to_original_function_in_test_mode(
     patch_destination, test_autologging_integration
 ):
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         return original("1", "2", "3")
 
@@ -400,7 +360,7 @@ def test_safe_patch_validates_arguments_to_original_function_in_test_mode(
             wraps=autologging_utils.safety._validate_args,
         ) as validate_mock,
     ):
-        _run_sync_or_sync(patch_destination.fn, "a", "b", "c")
+        run_sync_or_sync(patch_destination.fn, "a", "b", "c")
 
     assert validate_mock.call_count == 1
 
@@ -411,13 +371,13 @@ def test_safe_patch_throws_when_autologging_runs_are_leaked_in_test_mode(
 ):
     assert autologging_utils.is_testing()
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def leak_run_patch_impl(original, *args, **kwargs):
         mlflow.start_run(nested=True)
 
     safe_patch(test_autologging_integration, patch_destination, "fn", leak_run_patch_impl)
     with pytest.raises(AssertionError, match="leaked an active run"):
-        _run_sync_or_sync(patch_destination.fn)
+        run_sync_or_sync(patch_destination.fn)
 
     # End the leaked run
     mlflow.end_run()
@@ -437,12 +397,12 @@ def test_safe_patch_does_not_throw_when_autologging_runs_are_leaked_in_standard_
 ):
     assert not autologging_utils.is_testing()
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def leak_run_patch_impl(original, *args, **kwargs):
         mlflow.start_run(nested=True)
 
     safe_patch(test_autologging_integration, patch_destination, "fn", leak_run_patch_impl)
-    _run_sync_or_sync(patch_destination.fn)
+    run_sync_or_sync(patch_destination.fn)
     assert mlflow.active_run()
 
     # End the leaked run
@@ -477,7 +437,7 @@ def test_safe_patch_validates_autologging_runs_when_necessary_in_test_mode(
         with pytest.raises(
             AssertionError, match="failed to set autologging tag with expected value"
         ):
-            _run_sync_or_sync(patch_destination.fn)
+            run_sync_or_sync(patch_destination.fn)
         assert validate_run_mock.call_count == 1
 
         validate_run_mock.reset_mock()
@@ -485,7 +445,7 @@ def test_safe_patch_validates_autologging_runs_when_necessary_in_test_mode(
         with mlflow.start_run(nested=True):
             # If a user-generated run existed prior to the autologged training session, we expect
             # that safe patch will not attempt to validate it
-            _run_sync_or_sync(patch_destination.fn)
+            run_sync_or_sync(patch_destination.fn)
         assert not validate_run_mock.called
 
 
@@ -494,7 +454,7 @@ def test_safe_patch_does_not_validate_autologging_runs_in_standard_mode(
 ):
     assert not autologging_utils.is_testing()
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def no_tag_run_patch_impl(original, *args, **kwargs):
         with mlflow.start_run(nested=True):
             return original(*args, **kwargs)
@@ -505,12 +465,12 @@ def test_safe_patch_does_not_validate_autologging_runs_in_standard_mode(
         "mlflow.utils.autologging_utils.safety._validate_autologging_run",
         wraps=_validate_autologging_run,
     ) as validate_run_mock:
-        _run_sync_or_sync(patch_destination.fn)
+        run_sync_or_sync(patch_destination.fn)
 
         with mlflow.start_run(nested=True):
             # If a user-generated run existed prior to the autologged training session, we expect
             # that safe patch will not attempt to validate it
-            _run_sync_or_sync(patch_destination.fn)
+            run_sync_or_sync(patch_destination.fn)
 
         assert not validate_run_mock.called
 
@@ -521,7 +481,7 @@ def test_safe_patch_manages_run_if_specified_and_sets_expected_run_tags(
     client = MlflowClient()
     active_run = None
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal active_run
         active_run = mlflow.active_run()
@@ -541,7 +501,7 @@ def test_safe_patch_manages_run_if_specified_and_sets_expected_run_tags(
             test_autologging_integration, patch_destination, "fn", patch_impl, manage_run=True
         )
 
-    _run_sync_or_sync(patch_destination.fn)
+    run_sync_or_sync(patch_destination.fn)
     assert managed_run_mock.call_count == 1
     assert active_run is not None
     assert active_run.info.run_id is not None
@@ -555,7 +515,7 @@ def test_safe_patch_does_not_manage_run_if_unspecified(
 ):
     active_run = None
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal active_run
         active_run = mlflow.active_run()
@@ -567,7 +527,7 @@ def test_safe_patch_does_not_manage_run_if_unspecified(
         safe_patch(
             test_autologging_integration, patch_destination, "fn", patch_impl, manage_run=False
         )
-        _run_sync_or_sync(patch_destination.fn)
+        run_sync_or_sync(patch_destination.fn)
         assert managed_run_mock.call_count == 0
         assert active_run is None
 
@@ -575,7 +535,7 @@ def test_safe_patch_does_not_manage_run_if_unspecified(
 def test_safe_patch_preserves_signature_of_patched_function(
     patch_destination, test_autologging_integration
 ):
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def original(a, b, c=10, *, d=11):
         return 10
 
@@ -583,14 +543,14 @@ def test_safe_patch_preserves_signature_of_patched_function(
 
     patch_impl_called = False
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal patch_impl_called
         patch_impl_called = True
         return original(*args, **kwargs)
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
-    _run_sync_or_sync(patch_destination.fn, 1, 2)
+    run_sync_or_sync(patch_destination.fn, 1, 2)
     assert patch_impl_called
     assert inspect.signature(patch_destination.fn) == inspect.signature(original)
 
@@ -598,7 +558,7 @@ def test_safe_patch_preserves_signature_of_patched_function(
 def test_safe_patch_provides_original_function_with_expected_signature(
     patch_destination, test_autologging_integration
 ):
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def original(a, b, c=10, *, d=11):
         return 10
 
@@ -606,14 +566,14 @@ def test_safe_patch_provides_original_function_with_expected_signature(
 
     original_signature = False
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal original_signature
         original_signature = inspect.signature(original)
         return original(*args, **kwargs)
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
-    _run_sync_or_sync(patch_destination.fn, 1, 2)
+    run_sync_or_sync(patch_destination.fn, 1, 2)
     assert original_signature == inspect.signature(original)
 
 
@@ -625,7 +585,7 @@ def test_safe_patch_makes_expected_event_logging_calls_for_successful_patch_invo
     patch_session = None
     og_call_kwargs = {}
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal og_call_kwargs
         kwargs.update({"extra_func": picklable_exception_safe_function(lambda k: "foo")})
@@ -638,7 +598,7 @@ def test_safe_patch_makes_expected_event_logging_calls_for_successful_patch_invo
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
 
-    _run_sync_or_sync(patch_destination.fn, "a", 1, b=2)
+    run_sync_or_sync(patch_destination.fn, "a", 1, b=2)
     expected_order = ["patch_start", "original_start", "original_success", "patch_success"]
     assert [call.method for call in mock_event_logger.calls] == expected_order
     assert all(call.session == patch_session for call in mock_event_logger.calls)
@@ -662,7 +622,7 @@ def test_safe_patch_makes_expected_event_logging_calls_when_patch_impl_throws_an
 
     throw_location = None
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal throw_location
 
@@ -685,7 +645,7 @@ def test_safe_patch_makes_expected_event_logging_calls_when_patch_impl_throws_an
 
     for throw_location in ["before", "after"]:
         mock_event_logger.reset()
-        _run_sync_or_sync(patch_destination.fn)
+        run_sync_or_sync(patch_destination.fn)
         assert [call.method for call in mock_event_logger.calls] == expected_order
         patch_start, original_start, original_success, patch_error = mock_event_logger.calls
         assert patch_start.exception is None
@@ -704,7 +664,7 @@ def test_safe_patch_makes_expected_event_logging_calls_when_patch_impl_throws_an
 
     throw_location = None
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal throw_location
 
@@ -723,7 +683,7 @@ def test_safe_patch_makes_expected_event_logging_calls_when_patch_impl_throws_an
     for throw_location in ["before", "after"]:
         mock_event_logger.reset()
         with pytest.raises(Exception, match="throw from original"):
-            _run_sync_or_sync(patch_destination.throw_error_fn, original_err_to_raise)
+            run_sync_or_sync(patch_destination.throw_error_fn, original_err_to_raise)
         assert [call.method for call in mock_event_logger.calls] == expected_order
         patch_start, original_start, original_error = mock_event_logger.calls
         assert patch_start.exception is None
@@ -738,20 +698,20 @@ def test_safe_patch_makes_expected_event_logging_calls_when_original_function_th
 ):
     exc_to_raise = Exception("thrown from patch")
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def original(*args, **kwargs):
         raise exc_to_raise
 
     patch_destination.fn = original
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         original(*args, **kwargs)
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_impl)
 
     with pytest.raises(Exception, match="thrown from patch"):
-        _run_sync_or_sync(patch_destination.fn)
+        run_sync_or_sync(patch_destination.fn)
     expected_order = ["patch_start", "original_start", "original_error"]
     assert [call.method for call in mock_event_logger.calls] == expected_order
     patch_start, original_start, original_error = mock_event_logger.calls
@@ -767,7 +727,7 @@ def test_safe_patch_succeeds_when_event_logging_throws_in_standard_mode(
     patch_preamble_called = False
     patch_postamble_called = False
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_impl(original, *args, **kwargs):
         nonlocal patch_preamble_called
         patch_preamble_called = True
@@ -828,7 +788,7 @@ def test_safe_patch_succeeds_when_event_logging_throws_in_standard_mode(
 
     logger = ThrowingLogger()
     AutologgingEventLogger.set_logger(logger)
-    assert _run_sync_or_sync(patch_destination.fn) == PATCH_DESTINATION_FN_DEFAULT_RESULT
+    assert run_sync_or_sync(patch_destination.fn) == PATCH_DESTINATION_FN_DEFAULT_RESULT
     assert patch_preamble_called
     assert patch_postamble_called
     expected_calls = ["patch_start", "original_start", "original_success", "patch_success"]
@@ -1474,25 +1434,25 @@ def test_session_manager_creates_session_before_patch_executes(
 ):
     is_session_active = None
 
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def check_session_manager_status(original):
         nonlocal is_session_active
         is_session_active = _AutologgingSessionManager.active_session()
 
     safe_patch(test_autologging_integration, patch_destination, "fn", check_session_manager_status)
-    _run_sync_or_sync(patch_destination.fn)
+    run_sync_or_sync(patch_destination.fn)
     assert is_session_active is not None
 
 
 def test_session_manager_exits_session_after_patch_executes(
     patch_destination, test_autologging_integration
 ):
-    @_test_async(patch_destination.is_async)
+    @asyncify(patch_destination.is_async)
     def patch_fn(original):
         assert _AutologgingSessionManager.active_session() is not None
 
     safe_patch(test_autologging_integration, patch_destination, "fn", patch_fn)
-    _run_sync_or_sync(patch_destination.fn)
+    run_sync_or_sync(patch_destination.fn)
     assert _AutologgingSessionManager.active_session() is None
 
 
@@ -1513,7 +1473,7 @@ def test_original_fn_runs_if_patch_should_not_be_applied(patch_destination):
 
     @autologging_integration("test_respects_exclusive")
     def autolog(disable=False, exclusive=False, silent=False):
-        @_test_async(patch_destination.is_async)
+        @asyncify(patch_destination.is_async)
         def patch_impl(original, *args, **kwargs):
             nonlocal patch_impl_call_count
             patch_impl_call_count += 1
@@ -1523,7 +1483,7 @@ def test_original_fn_runs_if_patch_should_not_be_applied(patch_destination):
 
     autolog(exclusive=True)
     with mlflow.start_run():
-        _run_sync_or_sync(patch_destination.fn)
+        run_sync_or_sync(patch_destination.fn)
     assert patch_impl_call_count == 0
     assert patch_destination.fn_call_count == 1
 
@@ -1581,7 +1541,7 @@ def test_nested_call_autologging_disabled_when_top_level_call_autologging_failed
         "test_nested_call_autologging_disabled_when_top_level_call_autologging_failed"
     )
     def autolog(disable=False, exclusive=False, silent=False):
-        @_test_async(patch_destination.is_async)
+        @asyncify(patch_destination.is_async)
         def patch_impl(original, *args, **kwargs):
             nonlocal patch_impl_call_count
             patch_impl_call_count += 1
@@ -1605,7 +1565,7 @@ def test_nested_call_autologging_disabled_when_top_level_call_autologging_failed
         patch_impl_call_count = 0
         patch_destination.recurse_fn_call_count = 0
         with mlflow.start_run():
-            _run_sync_or_sync(patch_destination.recursive_fn, level=0, max_depth=max_depth)
+            run_sync_or_sync(patch_destination.recursive_fn, level=0, max_depth=max_depth)
         assert patch_impl_call_count == 1
         assert patch_destination.recurse_fn_call_count == max_depth + 1
